tests/locust/load-requirements.txt
msgpack-python
pyzmq
locustio
